+++ src/components/Card.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 import { useTranslations } from '../shared/hooks/useTranslations';
-import { copyToClipboard, getLinkHostname, openExternally } from '../shared/lib/utils/utils';
+import { copyToClipboard, getLinkHostname, openExternally } from '../shared/lib/utils/link';
 import Button from '../shared/ui/Button';
 import DropdownMenu, { DropdownOption } from '../shared/ui/DropdownMenu';
