+++ src/components/PasswordGeneratorModal.tsx
@@ -1,6 +1,7 @@
 import { FC, useState } from 'react';
 import { useTranslations } from '../shared/hooks/useTranslations';
-import { copyToClipboard, generatePassword } from '../shared/lib/utils/utils';
+import { generatePassword } from '../shared/lib/utils/generate';
+import { copyToClipboard } from '../shared/lib/utils/link';
 import Button from '../shared/ui/Button';
 import Modal from '../shared/ui/Modal';
