+++ src/db/useDatabase.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { uniqueId } from '../shared/lib/utils/utils';
+import { uniqueId } from '../shared/lib/utils/generate';
 import { WebsiteCard, WebsiteCardCreate } from '../types/types';
 import { DB_KEYS } from './keys';
