+++ src/shared/lib/utils/generate.ts
@@ -1,0 +1,26 @@
+export const uniqueId = (length = 16) =>
+  [...Array(length)].map(() => Math.random().toString(36).charAt(2)).join('');
+
+export const generatePassword = (length = 16) => {
+  const lowercase = 'abcdefghijklmnopqrstuvwxyz';
+  const uppercase = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+  const numbers = '0123456789';
+  const symbols = '!@#$%^&*()_+{}[]<>?/~|';
+
+  const allChars = lowercase + uppercase + numbers + symbols;
+  let password = '';
+
+  password += lowercase[Math.floor(Math.random() * lowercase.length)];
+  password += uppercase[Math.floor(Math.random() * uppercase.length)];
+  password += numbers[Math.floor(Math.random() * numbers.length)];
+  password += symbols[Math.floor(Math.random() * symbols.length)];
+
+  for (let i = 4; i < length; i++) {
+    password += allChars[Math.floor(Math.random() * allChars.length)];
+  }
+
+  return password
+    .split('')
+    .sort(() => Math.random() - 0.5)
+    .join('');
+};
